requirements.txt
flask==2.3.2
flask-cors==4.0.0
flask-sqlalchemy==3.0.5
gunicorn==21.2.0
openai==1.35.3
click==8.1.7
rich==13.7.1
requests==2.31.0
python-dotenv==1.0.0
pytest==7.4.0